+++ src/mbgl/text/placement.hpp
@@ -91,7 +91,7 @@ public:
     const RenderTile& tile;
     std::string sourceId;
     std::shared_ptr<FeatureIndex> featureIndex;
-    float sortKey;
+    double sortKey;
     size_t symbolInstanceStart;
     size_t symbolInstanceEnd;
 };
